requirements.txt
flask
flask-cors
requests
biopython
rdkit
meeko
numpy
scipy
